Added music & fixed functionality of Audio Manager
- Fixed slider issue (still need to set init value; sound plays at full volume but fader graphic starts at null)
- Added/ammended audio assets
- Revised jump sound mechanics in Player.cs, but reverted back to previous code after audiomanager issues repaired
