+++ Cake Islands/Assets/Scenes/Level5.unity
@@ -112,6 +112,75 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1001 &1750836794
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 5486725999043791426, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_Name
+      value: AudioManager (1)
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5486725999043791424, guid: 6f5669f3df69bfa4e8641f01967582c1,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 6f5669f3df69bfa4e8641f01967582c1, type: 3}
 --- !u!1 &1791896939
 GameObject:
   m_ObjectHideFlags: 0
