+++ Cake Islands/Assets/Scripts/FallingSpikeSound.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FallingSpikeSound : MonoBehaviour
+{
+    void SpikeFallAudio()
+    {
+        FindObjectOfType<AudioManager>().Play("FallingSpikes");
+    }
+}
