+++ Cake Islands/Assets/Scripts/Sound.cs
@@ -8,10 +8,14 @@ public class Sound
 
     public AudioClip clip;
 
+    public AudioMixerGroup mix;
+
     [Range(0f, 1f)]
     public float volume;
     [Range(.1f, 3f)]
     public float pitch;
+
+    public bool loop;
 
     [HideInInspector]
     public AudioSource source;
